♻️ Move preactions to their own file.
